requirements.txt
Pillow
requests
cairosvg
tkinter